ProductPage components hierarchy created
